requirements.txt
pandas
yfinance
pandas_ta
python-telegram-bot
python-dotenv
pytz
